Undo rollback of Engine/Plugins/Runtime/Harmonix/... changelist 32212777 Bug determined to be unrelated to this change.
[CL 32498589 by buzz burrowes in 5.4 branch]

+++ Engine/Plugins/Runtime/Harmonix/Source/HarmonixMetasound/Public/HarmonixMetasound/Components/MusicClockComponent.h
@@ -250,7 +250,7 @@ public:
 	float GetSongLengthBeats() const;
 
 	UFUNCTION(BlueprintPure, Category = "Song Data")
-	int32 GetSongLengthBars() const;
+	float GetSongLengthBars() const;
 
 	const FSongMaps& GetSongMaps() const;
